+++ src/fast_crc32.h
@@ -5,14 +5,14 @@
 
 #include <stdint.h>
 
-uint32_t crc32a_bitwise(uint8_t* buff, uint16_t len);
-uint32_t crc32b_bitwise(uint8_t* buff, uint16_t len);
-uint32_t crc32a_lookup_table(uint8_t* buff, uint32_t len);
-uint32_t crc32b_lookup_table(uint8_t* buff, uint16_t len);
-uint32_t crc32a_fast_bytewise(uint8_t* buff, uint32_t len);
-uint32_t crc32b_fast_bytewise(uint8_t* buff, uint32_t len);
-uint32_t crc32a_fast_nibblewise(uint8_t* buff, uint32_t len);
-uint32_t crc32b_fast_nibblewise(uint8_t* buff, uint32_t len);
+uint32_t crc32a_bitwise(uint8_t* buff, size_t len);
+uint32_t crc32b_bitwise(uint8_t* buff, size_t len);
+uint32_t crc32a_lookup_table(uint8_t* buff, size_t len);
+uint32_t crc32b_lookup_table(uint8_t* buff, size_t len);
+uint32_t crc32a_fast_bytewise(uint8_t* buff, size_t len);
+uint32_t crc32b_fast_bytewise(uint8_t* buff, size_t len);
+uint32_t crc32a_fast_nibblewise(uint8_t* buff, size_t len);
+uint32_t crc32b_fast_nibblewise(uint8_t* buff, size_t len);
 
 void crc32a_table_init(uint32_t *table);
 void crc32b_table_init(uint32_t *table);
